requirements.txt
langchain
PyPDF2
python-dotenv
streamlit
faiss-cpu
streamlit-extras
langchain_community
InstructorEmbedding
sentence_transformers
fireworks-ai